feat: write tests for mainPage, correct styles
