+++ src/components/result-component/result-component.test.tsx
@@ -59,7 +59,7 @@ describe('<Results />', () => {
     });
   });
 
-  it('Validate that clicking on a card opens a detailed card component', async () => {
+  it('Validate that clicking on a card opens a detailed card component', () => {
     renderComponent(mockDataTest);
     const card = screen.getAllByRole('card');
     fireEvent.click(card[0]);
