+++ src/pages/NotFoundPage/NotFoundPage.test.tsx
@@ -7,7 +7,7 @@ describe('<NotFoundPage />', () => {
     expect(<NotFoundPage />).toBeDefined();
   });
 
-  it('it renders the not found page', () => {
+  it('Ensure that the 404 page is displayed when navigating to an invalid route', () => {
     render(
       <MemoryRouter initialEntries={['/foo']}>
         <NotFoundPage />
